requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.54
boto3-stubs==1.42.54
botocore==1.42.54
botocore-stubs==1.42.41
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
librt==0.8.1
Mako==1.3.10
MarkupSafe==3.0.3
mypy==1.19.1
mypy-boto3-cloudformation==1.42.3
mypy-boto3-dynamodb==1.42.41
mypy-boto3-ec2==1.42.51
mypy-boto3-lambda==1.42.37
mypy-boto3-rds==1.42.51
mypy-boto3-s3==1.42.37
mypy-boto3-sqs==1.42.3
mypy_extensions==1.1.0
pathspec==1.0.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
ruff==0.15.2
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
types-awscrt==0.31.2
types-s3transfer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
